chor: reorganize folders
