+++ packages/frontend/src/features/ui/theme/ModeToggle.tsx
@@ -1,13 +1,13 @@
 import { Moon, Sun } from "lucide-react";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/features/ui/components/button";
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { useTheme } from "@/components/ThemeProvider";
+} from "@/features/ui/components/dropdown-menu";
+import { useTheme } from "@/features/ui/theme/ThemeProvider";
 
 export function ModeToggle() {
   const { setTheme } = useTheme();
