+++ packages/frontend/src/pages/App.tsx
@@ -1,9 +1,9 @@
 import { ApolloProvider } from "@apollo/client";
-import { AuthProvider } from "./AuthContext/AuthContext";
-import { client } from "./client";
+import { AuthProvider } from "@/features/auth/AuthContext/AuthContext";
+import { client } from "@/features/api/client";
 import Root from "./Root";
-import "./global.css";
-import { ThemeProvider } from "./components/ThemeProvider";
+import "@/global.css";
+import { ThemeProvider } from "../features/ui/theme/ThemeProvider";
 
 const App = () => {
   return (
